+++ configs/node.js
@@ -1,0 +1,11 @@
+module.exports = {
+  env: {
+    node: true
+  },
+  extends: [
+    'plugin:bpmn-io/recommended'
+  ],
+  parserOptions: {
+    ecmaVersion: 'latest'
+  }
+};